refactor:开源版，研发商店组件logo、文档链接中域名在数据库硬编码，导致用户切换域名后无法查看图片 #3780
